chore: split ViewProjectDetails components into separate files
