+++ packages/grant-explorer/src/features/round/ViewProjectDetails/components/Detail.tsx
@@ -1,0 +1,13 @@
+import { renderToHTML } from "common";
+
+export function Detail(props: { text: string; testID: string }) {
+  return (
+    <p
+      dangerouslySetInnerHTML={{
+        __html: renderToHTML(props.text.replace(/\n/g, "\n\n")),
+      }}
+      className="text-blue-800 text-md prose prose-h1:text-lg prose-h2:text-base prose-h3:text-base prose-a:text-blue-600 max-w-full"
+      data-testid={props.testID}
+    />
+  );
+}
