+++ packages/grant-explorer/src/features/round/ViewProjectDetails/components/ProjectDetailsTab.tsx
@@ -1,0 +1,23 @@
+import { Box, Tab, Tabs } from "@chakra-ui/react";
+
+export function ProjectDetailsTabs(props: {
+  tabs: string[];
+  onChange?: (tabIndex: number) => void;
+  selected: number;
+}) {
+  return (
+    <Box className="" bottom={0.5}>
+      {props.tabs.length > 0 && (
+        <Tabs
+          display="flex"
+          onChange={props.onChange}
+          defaultIndex={props.selected}
+        >
+          {props.tabs.map((tab, index) => (
+            <Tab key={index}>{tab}</Tab>
+          ))}
+        </Tabs>
+      )}
+    </Box>
+  );
+}
