+++ packages/grant-explorer/src/features/round/ViewProjectDetails/components/index.ts
@@ -1,0 +1,8 @@
+export { ApplicationFormAnswers } from "./ApplicationFormAnswers";
+export { Detail } from "./Detail";
+export { ProjectDetailsTabs } from "./ProjectDetailsTab";
+export { ProjectLinks } from "./ProjectLinks";
+export { ProjectLogo } from "./ProjectLogo";
+export { Sidebar } from "./Sidebar";
+export { ProjectStats } from "./ProjectStats";
+export { Stat } from "./Stat";
