+++ packages/grant-explorer/src/features/round/ViewProjectDetails/hooks/useProjectDetailsParams.tsx
@@ -1,0 +1,7 @@
+import { useParams } from "common";
+
+export const useProjectDetailsParams = useParams<{
+  chainId: string;
+  roundId: string;
+  applicationId: string;
+}>;
